Move compatibility module into compat package.
