+++ distutils/tests/test_extension.py
@@ -6,7 +6,7 @@ from distutils.extension import Extension, read_setup_file
 
 import pytest
 
-from .py38compat import check_warnings
+from .compat.py38 import check_warnings
 
 
 class TestExtension:
