+++ distutils/tests/test_unixccompiler.py
@@ -12,7 +12,7 @@ from distutils.util import _clear_cached_macosx_ver
 import pytest
 
 from . import support
-from .py38compat import EnvironmentVarGuard
+from .compat.py38 import EnvironmentVarGuard
 
 
 @pytest.fixture(autouse=True)
